% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_ensemble.R
\name{as_tibble.epi_ensemble}
\alias{as_tibble.epi_ensemble}
\title{Convert an \code{epi_ensemble} to a tibble.}
\usage{
\method{as_tibble}{epi_ensemble}(x, ...)
}
\arguments{
\item{x}{An \code{epi_ensemble}.}

\item{...}{Unused.}
}
\value{
The \verb{$predictions} tibble.
}
\description{
Convert an \code{epi_ensemble} to a tibble.
}
