% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_recommend.R
\name{as_tibble.epi_recommendation}
\alias{as_tibble.epi_recommendation}
\title{Convert an \code{epi_recommendation} to a tibble.}
\usage{
\method{as_tibble}{epi_recommendation}(x, ...)
}
\arguments{
\item{x}{An \code{epi_recommendation}.}

\item{...}{Unused.}
}
\value{
The \verb{$recommendations} tibble.
}
\description{
Convert an \code{epi_recommendation} to a tibble.
}
