% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_smc.R
\name{calc_smc_mbg}
\alias{calc_smc_mbg}
\title{Prepare SMC Data for MBG Analysis}
\usage{
calc_smc_mbg(
  dhs_kr,
  gps_data,
  survey_vars = list(cluster = "v001", age = "hw1", smc_primary = "hml43", smc_alt =
    "ml13g"),
  gps_vars = list(cluster = "DHSCLUST", lat = "LATNUM", lon = "LONGNUM")
)
}
\arguments{
\item{dhs_kr}{DHS Children's Recode (KR) dataset.}

\item{gps_data}{DHS GPS dataset with cluster coordinates.}

\item{survey_vars}{Named list mapping DHS variable names.}

\item{gps_vars}{Named list for GPS variable mapping.}
}
\value{
A data.table with columns:
\itemize{
\item cluster_id: Cluster identifier
\item indicator: Number of children who received SMC
\item samplesize: Total number of children in analysis
\item x: Longitude
\item y: Latitude
}
}
\description{
Prepares cluster-level Seasonal Malaria Chemoprevention (SMC) receipt data
for MBG analysis. SMC coverage among children under 5.
}
\details{
Methodology: \url{https://github.com/ahadi-analytics/sntmethods/blob/master/inst/methods/smc_dhs.yml}

SMC variable availability varies by survey. Common variables include:
\itemize{
\item hml43: SMC in malaria season (DHS-7+)
\item ml13g: Received antimalarial for prevention
}

This function first checks which SMC-related variables are available
and uses the most appropriate one.
}
\examples{
\dontrun{
smc_mbg <- calc_smc_mbg(
  dhs_kr = kr_data,
  gps_data = gps_data
)
}

}
