% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_calc_case_management.R
\name{case_management_dictionary}
\alias{case_management_dictionary}
\title{Case Management Indicator Dictionary}
\usage{
case_management_dictionary()
}
\value{
Tibble with columns: indicator, indicator_code, indicator_title,
numerator_description, denominator_description, denominator_code.
}
\description{
Returns the dictionary of effective case management indicators.
}
\keyword{internal}
