% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmc_to_date}
\alias{cmc_to_date}
\title{Convert DHS Century Month Code to Date}
\usage{
cmc_to_date(cmc)
}
\arguments{
\item{cmc}{Numeric vector of CMC values.}
}
\value{
Vector of Date objects (mid-month: day 15).
}
\description{
Converts DHS Century Month Code (CMC) values to R Date objects.
CMC is a date coding system used by DHS where CMC = (year - 1900) * 12 + month.
}
\details{
The DHS Century Month Code (CMC) encodes dates as the number of months since
December 1899. The formula is: CMC = (year - 1900) * 12 + month.

For example:
\itemize{
\item CMC 1 = January 1900
\item CMC 1324 = January 2010
\item CMC 1404 = August 2016
}
}
\examples{
\dontrun{
cmc_to_date(1)     # 1900-01-15
cmc_to_date(1324)  # 2010-01-15
cmc_to_date(1404)  # 2016-08-15
}

}
