% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_preflight.R
\name{dhs_preflight_write}
\alias{dhs_preflight_write}
\title{Write the preflight to disk}
\usage{
dhs_preflight_write(preflight, path, country_iso3)
}
\arguments{
\item{preflight}{Output of \code{dhs_preflight()}.}

\item{path}{Output directory.}

\item{country_iso3}{Country code used in the file name.}
}
\description{
Write the preflight to disk
}
