% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_incidence.R
\name{.calc_n0_internal}
\alias{.calc_n0_internal}
\title{Calculate N0 (Crude Incidence) - Internal}
\usage{
.calc_n0_internal(df, scale_factor)
}
\arguments{
\item{df}{Data frame with standardized column names}

\item{scale_factor}{Denominator for incidence rate}
}
\value{
Data frame with n0_cases and n0_incidence columns added
}
\description{
Calculate N0 (Crude Incidence) - Internal
}
\keyword{internal}
