% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_incidence.R
\name{.determine_incidence_level}
\alias{.determine_incidence_level}
\title{Determine Highest Valid Incidence Level - Internal}
\usage{
.determine_incidence_level(df, levels)
}
\arguments{
\item{df}{Data frame with N0-N5 calculations}

\item{levels}{Character vector of requested levels}
}
\value{
Data frame with incidence_level column added
}
\description{
Determine Highest Valid Incidence Level - Internal
}
\keyword{internal}
