% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_incidence.R
\name{.round_incidence_output}
\alias{.round_incidence_output}
\title{Round Incidence Output Values - Internal}
\usage{
.round_incidence_output(output)
}
\arguments{
\item{output}{The output list from calc_incidence.}
}
\value{
The output list with rounded values.
}
\description{
Rounds case counts to whole numbers and incidence rates to 2 decimal places.
}
\keyword{internal}
