% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_calc_irs.R
\name{irs_dictionary}
\alias{irs_dictionary}
\title{IRS Indicator Dictionary}
\usage{
irs_dictionary()
}
\value{
Tibble with columns: indicator, indicator_code, indicator_title,
numerator_description, denominator_description, denominator_code.
}
\description{
Returns the dictionary of IRS indicators with metadata.
}
\examples{
irs_dictionary()

}
