% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_calc_malaria_dx.R
\name{malaria_dx_dictionary}
\alias{malaria_dx_dictionary}
\title{Malaria Diagnostic Testing Indicator Dictionary}
\usage{
malaria_dx_dictionary()
}
\value{
Tibble with columns: indicator, indicator_code, indicator_title,
numerator_description, denominator_description, denominator_code.
}
\description{
Returns the dictionary of malaria diagnostic testing indicators with
metadata.
}
\examples{
malaria_dx_dictionary()

}
