% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhs_calc_pfpr.R
\name{pfpr_dictionary}
\alias{pfpr_dictionary}
\title{PfPR Indicator Dictionary}
\usage{
pfpr_dictionary()
}
\value{
Tibble with columns: indicator, indicator_code, indicator_title,
outcome_var, numerator_description, denominator_description,
denominator_code, data_level.
}
\description{
Returns the dictionary of PfPR indicators.
}
\examples{
pfpr_dictionary()

}
