% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_incidence.R
\name{plot.snt_incidence}
\alias{plot.snt_incidence}
\title{Plot Method for SNT Incidence Objects}
\usage{
\method{plot}{snt_incidence}(x, level = NULL, by = "adm2", max_facets = 12, ...)
}
\arguments{
\item{x}{An object of class \code{snt_incidence}.}

\item{level}{Character. Which incidence level to plot
(default: highest available).}

\item{by}{Character. Grouping variable for plot
(default: "adm2"). Set to NULL for aggregate plot.}

\item{max_facets}{Numeric. Maximum number of facets to display
(default: 12).}

\item{...}{Additional arguments passed to ggplot2 functions.}
}
\value{
A ggplot2 object.
}
\description{
Creates a time series plot of incidence rates by level. If multiple
administrative units are present, will facet by adm2 (up to 12 units).
}
\keyword{internal}
