% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_incidence.R
\name{print.snt_incidence}
\alias{print.snt_incidence}
\title{Print Method for SNT Incidence Objects}
\usage{
\method{print}{snt_incidence}(x, ...)
}
\arguments{
\item{x}{An object of class \code{snt_incidence}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for SNT Incidence Objects
}
\keyword{internal}
