% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_u5mr.R
\name{u5mr_dictionary}
\alias{u5mr_dictionary}
\title{U5MR Indicator Dictionary}
\usage{
u5mr_dictionary()
}
\value{
Tibble with columns: indicator, indicator_code, indicator_title,
numerator_description, denominator_description, denominator_code.
}
\description{
Returns the dictionary of U5MR indicators with metadata.
}
\keyword{internal}
