% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhs_read}
\alias{dhs_read}
\title{Filter DHS Parquet Dataset (Arrow)}
\usage{
dhs_read(
  path,
  survey_id = NULL,
  file_type = NULL,
  country_code = NULL,
  survey_year = NULL,
  survey_type = NULL,
  verbose = TRUE
)
}
\arguments{
\item{path}{Root parquet directory}

\item{survey_id}{Optional survey ID (e.g., "KEKR8A")}

\item{file_type}{Optional file type (PR, IR, KR, GE)}

\item{country_code}{Optional two-letter DHS country code}

\item{survey_year}{Optional numeric survey year}

\item{survey_type}{Optional DHS survey type (e.g., "DHS", "MIS")}

\item{verbose}{Logical; print progress messages? Default \code{TRUE}.}
}
\value{
A tibble of filtered DHS records
}
\description{
Filter DHS Parquet Dataset (Arrow)
}
